+++ tests/conftest.py
@@ -1,5 +1,3 @@
 """Functions, classes and fixtures available across all tests."""
-import os
-# print(f"Current dir is {os.getcwd()}.")
 
 from rdsa_utils.test_utils import *